requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
sqlmodel==0.0.8
pytest==7.4.0
httpx==0.24.1
google-api-python-client==2.88.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6